requirements.txt
fastapi
uvicorn
passlib
bcrypt
tortoise-orm
aerich
email-validator
fastapi-jwt-auth
python-multipart